Added example of new implemented method for overriding labels translating them to hex in the bottom slider.
I made a lot of adjustments to the storyboard because the last slider was below the fold in some of the smaller iPhone sizes.

I also pushed them examples closer together and ended up replacing the last example (no labels) with overriden labels.

I know it's important functionality but trivial to use as it's just a designable and was running out of room and a bit lazy to try and fit another slider in the limited space.

There were a lot of warnings that changed other files in the project but they aren't relevant to the code actual code changes.